requirements.txt
antlr4-python3-runtime==4.13.2
matplotlib==3.9.0